refactor(runner): split mitm_addon.py into usage.py and body_utils.py
Pure code motion — no behavior changes. Reduces mitm_addon.py from 1052
to 554 lines by extracting two new modules:

- usage.py: proxy-side usage extraction (Anthropic SSE, JSON) and
  reporting (HTTP POST, retry, thread pool). Home for future per-call
  connector billing extractors.
- body_utils.py: decompression (gzip/br/zstd), UTF-8-safe truncation,
  encoding, header redaction, capture fields.

mitm_addon.py uses module-prefixed calls (usage.X() / body_utils.X())
so tests can patch target symbols in one place regardless of caller.
This avoids the Python mock-placement pitfall where patching the
importer's binding fails to intercept calls from within the source
module.

Existing tests continue to exercise the same 762 Python + 616 Rust
test cases; patch targets updated to reflect symbol locations.

Closes #9460
